fix transitioning bug by only transitioning newly-created UserExercise objects
We had a bug where if the user was proficient in the new model but not the old
model, and transitioned to the new model, their progress bar would be full but
they wouldn't be marked as proficient yet. This fixes it by force using the old
streak model for all existing UserExercise objects and only allow using the new
accuracy model for newly-created UserExercise objects.