First cut at content dashboard.
This just contains information on searches that are potentially failing our users (as measured byGoogle analyitics' "exit rate"). In the future, this can house other statistics that attempt to estimate what users try to look for but fail to find, in terms of content.

The data from analytics is fetched via Google's JS API's using AuthSubJS. This requires the user to have access to the google analytics account.