Make the user's exercise listing sorted
This changes it so that the sort order of the exercise listing for the student is sorted according to the full list of exercises. We never guaranteed any ordering so this is not a breaking API change (but a desirable one for our internal consumption, so we can document this from now on...)

Also some minor cleanup to logic. The contents of the results should be the same other than sort order, though theoretically we used to be returning exercises that were in "UserExerciseGraph" that were deleted from the master Exercise table. I feel like we should just not return those, though. In practice, it probably doesn't matter since nobody relies/cares about them.