Minor template additions for "api based" links on the profile page
This is a pseudo-cherry-pick of Tom's work from goals which enables the profile page to create profile links that make API calls instead of custom graph calls.

This is a purely additive change and does not change the existing behavior of the app. A follow-up change will use these templates